force float32

+++ gradunwarp/core/tests/test_regression.py
@@ -10,7 +10,7 @@ def test_siemens_B():
     siemens_coeffs = coeffs.get_coefficients('siemens', gradfile)
     R0 = siemens_coeffs.R0_m  * 1000
 
-    vec = np.linspace(-300, 300, 60)
+    vec = np.linspace(-300, 300, 60, dtype=np.float32)
     x, y ,z = utils.meshgrid(vec, vec, vec)
 
     bx = siemens_B(siemens_coeffs.alpha_x, siemens_coeffs.beta_x, x, y, z, R0)
